Move addMediaChoice to AbstractSelectListAdapter where it's used
