+++ collect_app/src/test/java/org/odk/collect/android/widgets/SelectOneWidgetTest.java
@@ -1,9 +1,38 @@
 package org.odk.collect.android.widgets;
 
-import androidx.annotation.NonNull;
+import android.app.Application;
 
+import androidx.annotation.NonNull;
+import androidx.core.util.Pair;
+import androidx.lifecycle.MutableLiveData;
+
+import org.javarosa.core.model.SelectChoice;
+import org.javarosa.core.reference.ReferenceManager;
+import org.javarosa.form.api.FormEntryCaption;
+import org.junit.Before;
+import org.junit.Rule;
+import org.junit.Test;
+import org.mockito.Mock;
+import org.mockito.junit.MockitoJUnit;
+import org.mockito.junit.MockitoRule;
+import org.odk.collect.android.analytics.Analytics;
+import org.odk.collect.android.audio.AudioButton;
+import org.odk.collect.android.audio.AudioHelper;
+import org.odk.collect.android.audio.Clip;
+import org.odk.collect.android.formentry.media.AudioHelperFactory;
 import org.odk.collect.android.formentry.questions.QuestionDetails;
+import org.odk.collect.android.injection.config.AppDependencyModule;
+import org.odk.collect.android.support.MockFormEntryPromptBuilder;
+import org.odk.collect.android.support.RobolectricHelpers;
 import org.odk.collect.android.widgets.base.GeneralSelectOneWidgetTest;
+
+import static java.util.Arrays.asList;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+import static org.odk.collect.android.support.Helpers.createMockReference;
+import static org.odk.collect.android.support.RobolectricHelpers.populateRecyclerView;
 
 /**
  * @author James Knight
@@ -15,5 +44,86 @@ public class SelectOneWidgetTest extends GeneralSelectOneWidgetTest<AbstractSele
     @Override
     public SelectOneWidget createWidget() {
         return new SelectOneWidget(activity, new QuestionDetails(formEntryPrompt, "formAnalyticsID"), false);
+    }
+
+    @Rule
+    public MockitoRule rule = MockitoJUnit.rule();
+
+    @Mock
+    public ReferenceManager referenceManager;
+
+    @Mock
+    public AudioHelper audioHelper;
+
+    @Mock
+    public Analytics analytics;
+
+    @Before
+    public void setup() {
+        overrideDependencyModule();
+        when(audioHelper.setAudio(any(AudioButton.class), any())).thenReturn(new MutableLiveData<>());
+    }
+
+    @Test
+    public void whenChoicesHaveAudio_audioButtonUsesIndexAsClipID() throws Exception {
+        formEntryPrompt = new MockFormEntryPromptBuilder()
+                .withIndex("i am index")
+                .withSelectChoices(asList(
+                        new SelectChoice("1", "1"),
+                        new SelectChoice("2", "2")
+                ))
+                .withSpecialFormSelectChoiceText(asList(
+                        new Pair<>(FormEntryCaption.TEXT_FORM_AUDIO, "file://blah1.mp3"),
+                        new Pair<>(FormEntryCaption.TEXT_FORM_AUDIO, "file://blah2.mp3")
+                ))
+                .build();
+
+        String reference1 = createMockReference(referenceManager, "file://blah1.mp3");
+        String reference2 = createMockReference(referenceManager, "file://blah2.mp3");
+
+        populateRecyclerView(getActualWidget());
+        verify(audioHelper).setAudio(any(AudioButton.class), eq(new Clip("i am index 0", reference1)));
+        verify(audioHelper).setAudio(any(AudioButton.class), eq(new Clip("i am index 1", reference2)));
+    }
+
+    @Test
+    public void whenChoicesHaveAudio_logsAudioChoiceEvent() throws Exception {
+        createMockReference(referenceManager, "file://blah1.mp3");
+        createMockReference(referenceManager, "file://blah2.mp3");
+
+        formEntryPrompt = new MockFormEntryPromptBuilder()
+                .withIndex("i am index")
+                .withSelectChoices(asList(
+                        new SelectChoice("1", "1"),
+                        new SelectChoice("2", "2")
+                ))
+                .withSpecialFormSelectChoiceText(asList(
+                        new Pair<>(FormEntryCaption.TEXT_FORM_AUDIO, "file://blah1.mp3"),
+                        new Pair<>(FormEntryCaption.TEXT_FORM_AUDIO, "file://blah2.mp3")
+                ))
+                .build();
+
+        populateRecyclerView(getActualWidget());
+        verify(analytics).logEvent("Prompt", "AudioChoice", "formAnalyticsID");
+    }
+
+    private void overrideDependencyModule() {
+        RobolectricHelpers.overrideAppDependencyModule(new AppDependencyModule() {
+
+            @Override
+            public ReferenceManager providesReferenceManager() {
+                return referenceManager;
+            }
+
+            @Override
+            public AudioHelperFactory providesAudioHelperFactory() {
+                return context -> audioHelper;
+            }
+
+            @Override
+            public Analytics providesAnalytics(Application application) {
+                return analytics;
+            }
+        });
     }
 }
